Move properties into a mixin module (verbatim move)
